backend_api/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
python-multipart
python-dotenv
pydantic-settings
timm